requirements.txt
# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
pydicom==2.4.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Deep Learning Utilities
tqdm==4.66.1
tensorboard==2.14.1

# Metrics and Evaluation
scikit-image==0.21.0

# Streamlit App
streamlit==1.28.0
streamlit-option-menu==0.3.6

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Medical Imaging
nibabel==5.1.0

# For Grad-CAM
grad-cam==1.4.8

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.25.2
